# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.SysUser:
    properties:
      ID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      user_name:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      msg:
        type: string
      requestId:
        description: 数据集
        type: string
      status:
        type: string
    type: object
  router.UpdateUserInfoRequest:
    properties:
      email:
        type: string
      phone:
        type: string
    type: object
  router.UpdateUserPasswordRequest:
    properties:
      new_password:
        type: string
      password:
        description: UserId      int64  `json:"user_id"`
        type: string
    type: object
  service.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: string
      user_name:
        type: string
    type: object
  service.GetUsersResponse:
    properties:
      count:
        type: integer
      page_num:
        type: integer
      page_size:
        type: integer
      sysUsers:
        items:
          $ref: '#/definitions/models.SysUser'
        type: array
    type: object
  service.LoginRequest:
    properties:
      password:
        type: string
      user_name:
        type: string
    type: object
  service.ResetUserPasswordRequest:
    properties:
      new_password:
        type: string
      user_id:
        type: integer
    type: object
  service.UpdateUsersInfoRequest:
    properties:
      email:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact:
    email: 40010355@qq.com
    name: 刘炜
    url: https://blog.csdn.net/xingzuo_1840
  termsOfService: http://127.0.0.1
  title: crow-qin
  version: "1.0"
paths:
  /api/v1/admin/user:
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 填写用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 创建用户
      tags:
      - admin
  /api/v1/admin/user-password:
    put:
      consumes:
      - application/json
      description: admin重置用户密码
      parameters:
      - description: 填写用户Id和xin密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/service.ResetUserPasswordRequest'
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: admin重置用户密码
      tags:
      - admin
  /api/v1/admin/users:
    get:
      description: 获取用户信息
      parameters:
      - description: 用户角色
        in: query
        name: role
        type: string
      - description: 用户名（可模糊查询）
        in: query
        name: user_name
        type: string
      - description: 页数
        in: query
        name: page_num
        type: string
      - description: 每页行数
        in: query
        name: page_size
        type: string
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.GetUsersResponse'
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 获取用户信息
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: 修改用户信息
      parameters:
      - description: user_id必须，其他可选
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUsersInfoRequest'
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 修改用户信息
      tags:
      - admin
  /api/v1/admin/users/{uuid}:
    delete:
      description: 删除指定用户
      parameters:
      - description: 用户id
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 删除指定用户
      tags:
      - admin
    get:
      description: 获取指定用户信息
      parameters:
      - description: 用户id
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SysUser'
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 获取指定用户信息
      tags:
      - admin
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 用户登录，并获取token
      parameters:
      - description: 用户名，用户密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/service.LoginRequest'
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 用户登录
      tags:
      - system
  /api/v1/login/refresh:
    post:
      consumes:
      - application/json
      description: 刷新用户token
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 刷新token
      tags:
      - system
  /api/v1/logout:
    post:
      consumes:
      - application/json
      description: 用户登出
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 用户登出
      tags:
      - system
  /api/v1/user/info:
    get:
      description: 用户获取本人信息
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SysUser'
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 用户获取本人信息
      tags:
      - user
    put:
      consumes:
      - application/json
      description: 用户修改自己信息
      parameters:
      - description: 选项可选
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/router.UpdateUserInfoRequest'
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 用户修改自己信息
      tags:
      - user
  /api/v1/user/password:
    put:
      consumes:
      - application/json
      description: 用户修改自己的密码
      parameters:
      - description: 用户的新密码和旧密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/router.UpdateUserPasswordRequest'
      responses:
        "200":
          description: '{"requestId": "string","code": 200,"msg": "ok","data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: '{"requestId": "string","code": 500,"msg": "string","status":
            "error","data": null}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 用户修改密码
      tags:
      - user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"